requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain
langchain-community
langchain-huggingface
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
groq
